1st skeleton of the project
The file "thingPropertoes.h" is important as it contains variables defined in set (and not in code).
I added some constants representing the amount of water and light required, and left comments all along the way.
The alerts for now are only to the monitor and should be to the blynk app.
For now, I used delay() to make it sleep for 1 hour and then sample again, but not sure it's the best approach. Using some time libraries to work around it seems like a big headache.